refactor: move toolbar to workspace/toolbar
