+++ packages/base/src/index.ts
@@ -8,7 +8,7 @@ export * from './menus';
 export * from './panelview';
 export * from './store';
 export * from './features/stac-browser';
-export * from './toolbar';
+export * from './workspace/toolbar';
 export * from './tools';
 export * from './types';
 export * from './widget';
